service: move integration tests to separate files
